# perf: optimize search from 2000ms to 141ms with lazy loading