… tests/lazy_loading_robustness_tests.rs